requirements.txt
Django==4.2.13
djangorestframework==3.15.2
isort==5.13.2
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0

